move level loading related assets into /assets
