+++ website/levels/viewer/index.html
@@ -15,7 +15,6 @@
         <link rel="stylesheet" href="styles.css">
     </head>
     <body>
-        <script type="module" src="scripts/shaders.js"></script>
         <script type="module" src="scripts/level_viewer.js"></script>
 
         <div class="popup" id="popup">
@@ -39,7 +38,7 @@
         <div class="popup" id="popup-2">
 			<div class="popup-inner">
                 <div id="popup-thumbnail">
-                    <img src="./textures/preview_image_placeholder.png" id="popup-report-image">
+                    <img src="../../src/assets/preview_image_placeholder.png" id="popup-report-image">
                     <button id="report-set-image"><span>Set Image</span></button>
                 </div>
                 <br>
