+++ website/src/assets/models/cone.gltf
