+++ website/src/assets/models/cylinder.gltf
