+++ website/src/assets/models/pyramid.gltf
