+++ website/src/assets/textures/default_colored.png
