+++ website/src/assets/textures/grabbable.png
